requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.26.164
botocore==1.29.164
Django==4.2.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
